requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
django-money==3.0.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
phonenumberslite==8.12.56
psycopg2-binary==2.9.3
py-moneyed==2.0
pytz==2022.2.1
sqlparse==0.4.3
typing_extensions==4.3.0
